+++ src/main/java/net/silentchaos512/gear/traits/conditions/GearTypeTraitCondition.java
@@ -44,7 +44,7 @@ public class GearTypeTraitCondition implements ITraitCondition {
     @Override
     public boolean matches(ItemStack gear, PartType partType, Collection<MaterialInstance> materials, ITrait trait) {
         GearType type = GearHelper.getType(gear);
-        return type != null && type.matches(this.gearType);
+        return type == null || type.matches(this.gearType);
     }
 
     public static class Serializer implements ITraitConditionSerializer<GearTypeTraitCondition> {