+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - 'server' config is now 'common' and will appear in the config folder, like most configs do
 ### Fixed
 - Shears not losing durability on leaves [#188]
+- Parts not showing some conditional traits [#186]
 
 ## [1.10.2] - 2020-07-23
 ### Added